+++ gems/aws-sdk-core/lib/aws-sdk-core/rpc_v2/error_handler.rb
@@ -27,7 +27,7 @@ module Aws
       end
 
       def extract_error(body, context)
-        data = Cbor.decode(body)
+        data = RpcV2.decode(body)
         code = error_code(data, context)
         message = data['message']
         data = parse_error_data(context, body, code)